add Router_error.cpp
