+++ http/Router.cpp
@@ -116,60 +116,3 @@ void Router::handleDelete() {
 		std::cerr << "Error: " << e.what() << std::endl;
 	}
 }
-
-bool Router::isHeaderEnd() {
-	return request.isHeaderEnd();
-}
-
-bool Router::isRequestEnd() {
-	return request.isRequestEnd();
-}
-
-bool Router::getHaveResponse(void) const {
-	return haveResponse;
-}
-
-const std::string& Router::getResponse(void) const {
-	return response.getResponseStr();
-}
-
-void Router::addRequest(const std::string &request) {
-	this->request.addRequest(request);
-}
-
-void Router::parseHeader(void) {
-	request.parseHeader();
-}
-
-void Router::parseBody(void) {
-	request.parseBody();
-}
-
-void Router::setResponse(const std::string &src) {
-	response.setResponse(src);
-}
-
-void Router::readCGI(void) {
-	response.readCGI();
-}
-
-void Router::writeCGI(const intptr_t fdBufferSize) {
-	response.writeCGI(fdBufferSize);
-}
-
-void Router::disconnectCGI(void) {
-	response.disconnectCGI();
-	haveResponse = true;
-}
-
-int Router::getWriteFd(void) const {
-	return response.getWriteFd();
-}
-
-int Router::getReadFd(void) const {
-	return response.getReadFd();
-}
-
-int Router::getRequestError() const {
-    return request.getError();
-}
