+++ http/Router_utils.cpp
@@ -6,7 +6,6 @@
 #include <iomanip>
 
 static const std::string	g_dir = "./document";
-static const std::string    g_error_dir = g_dir + "/error.html";
 static const std::string    post_txt = g_dir + "/posts.txt";
 static const std::string    index_html = g_dir + "/index.html";
 
@@ -76,15 +75,6 @@ void Router::readFile(const std::string& filePath, std::string& content) {
 
 const std::string& Router::getResponseStr(void) const {
 	return response.getResponseStr();
-}
-
-void Router::makeErrorPage(void) {
-    makeErrorResponse(404);
-    if (resourceExists(g_error_dir)) {
-        std::string data;
-        readFile(g_error_dir, data);
-        response.makeBody(data, data.length(), getMIME(g_error_dir));
-    }
 }
 
 std::string Router::URLDecode(const std::string &input) {
@@ -221,7 +211,6 @@ void Router::readAndModifyHTML(std::string& htmlResponse) {
     }
 }
 
-
 void Router::makeHTMLResponse(const std::string& htmlResponse) {
 	response.makeStatusLine("HTTP/1.1", "200", "OK");
 	response.makeBody(htmlResponse, htmlResponse.size(), "text/html");
@@ -240,63 +229,59 @@ std::string Router::readPosts() {
 	return postHtml;
 }
 
-void Router::makeErrorResponse(int statusCode) {
-    std::string reasonPhrase;
-    std::string body;
+bool Router::isHeaderEnd() {
+	return request.isHeaderEnd();
+}
 
-    switch (statusCode) {
-        case 400:
-            reasonPhrase = "Bad Request";
-            body = "The request could not be understood by the server due to malformed syntax.";
-            break;
-        case 401:
-            reasonPhrase = "Unauthorized";
-            body = "The request requires user authentication.";
-            break;
-        case 403:
-            reasonPhrase = "Forbidden";
-            body = "The server understood the request, but is refusing to fulfill it.";
-            break;
-        case 404:
-            reasonPhrase = "Not Found";
-            body = "The server has not found anything matching the Request-URI.";
-            break;
-        case 405:
-            reasonPhrase = "Method Not Allowed";
-            body = "The method specified in the Request-Line is not allowed for the resource identified by the Request-URI.";
-            break;
-        case 411:
-            reasonPhrase = "Length Required";
-            body = "The request did not specify the length of its content, which is required by the requested resource.";
-            break;
-        case 413:
-            reasonPhrase = "Request Entity Too Large";
-            body = "The server is refusing to process a request because the request entity is larger than the server is willing or able to process.";
-            break;
-        case 415:
-            reasonPhrase = "Unsupported Media Type";
-            body = "The server is refusing to service the request because the entity of the request is in a format not supported by the requested resource for the requested method.";
-            break;
-        case 500:
-            reasonPhrase = "Internal Server Error";
-            body = "The server encountered an unexpected condition which prevented it from fulfilling the request.";
-            break;
-        case 501:
-            reasonPhrase = "Not Implemented";
-            body = "The server does not support the functionality required to fulfill the request.";
-            break;
-        case 505:
-            reasonPhrase = "HTTP Version Not Supported";
-            body = "The server does not support, or refuses to support, the HTTP protocol version that was used in the request message.";
-            break;
-        default:
-            statusCode = 500;
-            reasonPhrase = "Internal Server Error";
-            body = "An unexpected error occurred.";
-            break;
-    }
+bool Router::isRequestEnd() {
+	return request.isRequestEnd();
+}
 
-    response.makeStatusLine("HTTP/1.1", std::to_string(statusCode), reasonPhrase);
-    response.makeBody(body, body.length(), "text/plain");
-    haveResponse = true;
+bool Router::getHaveResponse(void) const {
+	return haveResponse;
+}
+
+const std::string& Router::getResponse(void) const {
+	return response.getResponseStr();
+}
+
+void Router::addRequest(const std::string &request) {
+	this->request.addRequest(request);
+}
+
+void Router::parseHeader(void) {
+	request.parseHeader();
+}
+
+void Router::parseBody(void) {
+	request.parseBody();
+}
+
+void Router::setResponse(const std::string &src) {
+	response.setResponse(src);
+}
+
+void Router::readCGI(void) {
+	response.readCGI();
+}
+
+void Router::writeCGI(const intptr_t fdBufferSize) {
+	response.writeCGI(fdBufferSize);
+}
+
+void Router::disconnectCGI(void) {
+	response.disconnectCGI();
+	haveResponse = true;
+}
+
+int Router::getWriteFd(void) const {
+	return response.getWriteFd();
+}
+
+int Router::getReadFd(void) const {
+	return response.getReadFd();
+}
+
+int Router::getRequestError() const {
+    return request.getError();
 }
